added integral metrics with sample rates

+++ src/System.Metrics/Endpoint.cs
@@ -6,7 +6,7 @@ namespace System.Metrics
 
         void Record<TMetric>(string metric, double value) where TMetric : IAllowsDouble;
 
-        void Record<TMetric>(string metric, double value, int sampleRate) where TMetric : IAllowsInteger, IAllowsSampleRate;
+        void Record<TMetric>(string metric, int value, double sampleRate) where TMetric : IAllowsInteger, IAllowsSampleRate;
 
         void Record<TMetric>(string metric, double value, double sampleRate) where TMetric : IAllowsDouble, IAllowsSampleRate;
